Syntax and build fixes. Now builds

+++ include/embla_hardware/embla_hardware.h
@@ -35,6 +35,7 @@
 // #include "husky_msgs/HuskyStatus.h"
 #include <string>
 #include "roboclaw/roboclaw_driver.h"
+#include <map>
 
 namespace embla_hardware
 {
@@ -64,8 +65,8 @@ namespace embla_hardware
 
 		void registerControlInterfaces();
 
-		double encoderPulsesToAngular( const double &encoder ) const;
-		double angularToEncoderPulses( const double &angle ) const;
+		double encoderPulsesToAngular( const int &encoder ) const;
+		int angularToEncoderPulses( const double &angle ) const;
 
 		void limitDifferentialSpeed( double &travel_speed_left, double &travel_speed_right, double maxSpeed ) const;
